# chore: optimize generation speed