--- a/frontend/appflowy_flutter/packages/appflowy_result/pubspec.yaml
+++ b/frontend/appflowy_flutter/packages/appflowy_result/pubspec.yaml
@@ -4,16 +4,10 @@ version: 0.0.1
 homepage:
 
 environment:
-  sdk: '>=3.3.0 <4.0.0'
+  sdk: ">=3.3.0 <4.0.0"
   flutter: ">=1.17.0"
 
-dependencies:
-  flutter:
-    sdk: flutter
-
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
   flutter_lints: ^3.0.0
 
 # For information on the generic Dart part of this file, see the