--- a/frontend/appflowy_flutter/packages/flowy_infra/pubspec.yaml
+++ b/frontend/appflowy_flutter/packages/flowy_infra/pubspec.yaml
@@ -10,11 +10,9 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  flutter_svg: ^2.0.2
   json_annotation: ^4.7.0
   path_provider: ^2.0.15
   path: ^1.8.2
-  textstyle_extensions: "2.0.0-nullsafety"
   time: ">=2.0.0"
   uuid: ">=2.2.2"
   bloc: ^8.1.2
@@ -23,8 +21,6 @@ dependencies:
   file: ^7.0.0
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
   build_runner: ^2.2.0
   flutter_lints: ^3.0.1
   freezed: ^2.4.7