--- a/frontend/appflowy_flutter/packages/flowy_svg/analysis_options.yaml
+++ b/frontend/appflowy_flutter/packages/flowy_svg/analysis_options.yaml
@@ -1,3 +1,1 @@
-include: package:very_good_analysis/analysis_options.yaml
-
 linter: