--- a/frontend/scripts/code_generation/flowy_icons/generate_flowy_icons.sh
+++ b/frontend/scripts/code_generation/flowy_icons/generate_flowy_icons.sh
@@ -24,7 +24,7 @@ while [[ $# -gt 0 ]]; do
         verbose=true
         shift
         ;;
-    --include-packages)
+    --exclude-packages)
         shift
         ;;
     *)