--- a/frontend/scripts/code_generation/freezed/generate_freezed.sh
+++ b/frontend/scripts/code_generation/freezed/generate_freezed.sh
@@ -7,6 +7,33 @@ start_time=$(date +%s)
 skip_pub_get=false
 skip_pub_packages_get=false
 verbose=false
+include_packages=false
+
+# Parse command line arguments
+while [[ $# -gt 0 ]]; do
+  case "$1" in
+  --skip-pub-get)
+    skip_pub_get=true
+    shift
+    ;;
+  --skip-pub-packages-get)
+    skip_pub_packages_get=true
+    shift
+    ;;
+  --verbose)
+    verbose=true
+    shift
+    ;;
+  --include-packages)
+    include_packages=true
+    shift
+    ;;
+  *)
+    echo "Unknown option: $1"
+    exit 1
+    ;;
+  esac
+done
 
 # Store the current working directory
 original_dir=$(pwd)
@@ -17,33 +44,58 @@ cd "$(dirname "$0")"
 cd ../../../appflowy_flutter
 
 # Navigate to the appflowy_flutter directory and generate files
-echo "Generating files for appflowy_flutter"
+echo "🧊 Start generating freezed files (AppFlowy)."
 
-flutter packages pub get >/dev/null 2>&1
-
-dart run build_runner build -d
-echo "Done generating files for appflowy_flutter"
-
-echo "Generating files for packages"
-cd packages
-for d in */; do
-  # Navigate into the subdirectory
-  cd "$d"
-
-  # Check if the subdirectory contains a pubspec.yaml file
-  if [ -f "pubspec.yaml" ]; then
-    echo "Generating freezed files in $d..."
-    echo "Please wait while we clean the project and fetch the dependencies."
-    flutter packages pub get >/dev/null 2>&1
-    dart run build_runner build -d
-    echo "Done running build command in $d"
+if [ "$skip_pub_packages_get" = false ]; then
+  if [ "$verbose" = true ]; then
+    flutter packages pub get
   else
-    echo "No pubspec.yaml found in $d, it can\'t be a Dart project. Skipping."
+    flutter packages pub get >/dev/null 2>&1
   fi
+fi
 
-  # Navigate back to the packages directory
-  cd ..
-done
+if [ "verbose" = true ]; then
+  dart run build_runner build -d
+else
+  dart run build_runner build >/dev/null 2>&1
+fi
+
+if [ "$include_packages" = true ]; then
+  # Navigate to the packages directory
+  cd packages
+  for d in */; do
+    # Navigate into the subdirectory
+    cd "$d"
+
+    # Check if the pubspec.yaml file exists and contains the freezed dependency
+    if [ -f "pubspec.yaml" ] && grep -q "freezed" pubspec.yaml; then
+      echo "🧊 Start generating freezed files ($d)."
+      if [ "$skip_pub_packages_get" = false ]; then
+        if [ "$verbose" = true ]; then
+          flutter packages pub get
+        else
+          flutter packages pub get >/dev/null 2>&1
+        fi
+      fi
+      if [ "verbose" = true ]; then
+        dart run build_runner build -d
+      else
+        dart run build_runner build >/dev/null 2>&1
+      fi
+      echo "🧊 Done generating freezed files ($d)."
+    fi
+
+    # Navigate back to the packages directory
+    cd ..
+  done
+fi
 
 # Return to the original directory
 cd "$original_dir"
+
+echo "🧊 Done generating freezed files."
+
+# echo the cost time
+end_time=$(date +%s)
+cost_time=$((end_time - start_time))
+echo "🧊 Freezed files generation cost $cost_time seconds."