--- a/frontend/scripts/code_generation/language_files/generate_language_files.sh
+++ b/frontend/scripts/code_generation/language_files/generate_language_files.sh
@@ -25,7 +25,7 @@ while [[ $# -gt 0 ]]; do
         verbose=true
         shift
         ;;
-    --include-packages)
+    --exclude-packages)
         shift
         ;;
     *)