<!---
Thank you for submitting a pull request to AppFlowy. The team will dedicate their best efforts to reviewing and approving your pull request. If you have any questions about the project or feedback for us, please join our [Discord](https://discord.gg/wdjWUXXhtw).
-->

<!---
If your pull request adds a new feature, please drag and drop a video into this section to showcase what you've done! If not, you may delete this section.
-->

### Feature Preview

The cost time for clean build reduced from 160s to 110s.

#### Before
<img width="876" alt="Screenshot 2024-08-12 at 09 16 21" src="https://github.com/user-attachments/assets/0960f308-7166-492a-98de-98b010031075">

#### After
<img width="618" alt="Screenshot 2024-08-12 at 09 16 26" src="https://github.com/user-attachments/assets/3557e13f-a0d3-40d8-a040-5edc51e426af">

The cost time for incremental build reduced from 130s to 43s. (change one freezed file, one svg and one translation)

#### Before
<img width="486" alt="Screenshot 2024-08-12 at 09 22 43" src="https://github.com/user-attachments/assets/cae8d3cd-10d1-4495-94c6-ba7411c3e0bd">

#### After
<img width="785" alt="Screenshot 2024-08-12 at 09 30 26" src="https://github.com/user-attachments/assets/9d48097a-5159-4269-a8e7-9a334fa0a4e7">

<!---
List at least one issue here that this PR addresses. If it fixes the issue, please use the [fixes](https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/using-keywords-in-issues-and-pull-requests) keyword to close the issue. For example:
fixes https://github.com/AppFlowy-IO/AppFlowy/pull/2106
-->

---

<!---
Before you mark this PR ready for review, run through this checklist!
-->

#### PR Checklist

- [x] My code adheres to [AppFlowy's Conventions](https://docs.appflowy.io/docs/documentation/software-contributions/conventions)
- [ ] I've listed at least one issue that this PR fixes in the description above.
- [ ] I've added a test(s) to validate changes in this PR, or this PR only contains semantic changes.
- [ ] All existing tests are passing.